src: refactor non-raw versions of the APIs wrt error & result
Change the non-raw APIs to check the exit code to match that of
the underlying zoau commands, and throw an exception on error
for those that didn't.

Also change delete and delete_members to process the option
"force" (currently not in Python's zoau APIs) so that, if present
and the dataset or member doesn't exist, then don't consider it
an error (like rm -f).